+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "com.hermosotech.myapplication"
+        applicationId "com.hermosotech.filmjoy"
         minSdk 21
         targetSdk 32
         versionCode 1
