+++ app/src/androidTest/java/com/hermosotech/filmjoy/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.hermosotech.myapplication
+package com.hermosotech.filmjoy
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.hermosotech.myapplication", appContext.packageName)
+        assertEquals("com.hermosotech.filmjoy", appContext.packageName)
     }
 }
