+++ app/src/main/java/com/hermosotech/filmjoy/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.hermosotech.myapplication
+package com.hermosotech.filmjoy
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
